Moved package permission configuration checks to separate module
