+++ lib/spack/spack/test/concretize_preferences.py
@@ -6,8 +6,8 @@
 import pytest
 import stat
 
-import spack.package_prefs
 import spack.repo
+import spack.package_perms as spp
 import spack.util.spack_yaml as syaml
 from spack.config import ConfigScope, ConfigError
 from spack.spec import Spec
@@ -246,42 +246,42 @@ mpi:
 
         # Test inheriting from 'all'
         spec = Spec('zmpi')
-        perms = spack.package_prefs.get_package_permissions(spec)
+        perms = spp.get_package_permissions(spec)
         assert perms == stat.S_IRWXU | stat.S_IRWXG
 
-        dir_perms = spack.package_prefs.get_package_dir_permissions(spec)
+        dir_perms = spp.get_package_dir_permissions(spec)
         assert dir_perms == stat.S_IRWXU | stat.S_IRWXG | stat.S_ISGID
 
-        group = spack.package_prefs.get_package_group(spec)
+        group = spp.get_package_group(spec)
         assert group == 'all'
 
     def test_config_permissions_from_package(self, configure_permissions):
         # Test overriding 'all'
         spec = Spec('mpich')
-        perms = spack.package_prefs.get_package_permissions(spec)
+        perms = spp.get_package_permissions(spec)
         assert perms == stat.S_IRWXU
 
-        dir_perms = spack.package_prefs.get_package_dir_permissions(spec)
+        dir_perms = spp.get_package_dir_permissions(spec)
         assert dir_perms == stat.S_IRWXU
 
-        group = spack.package_prefs.get_package_group(spec)
+        group = spp.get_package_group(spec)
         assert group == 'all'
 
     def test_config_permissions_differ_read_write(self, configure_permissions):
         # Test overriding group from 'all' and different readable/writable
         spec = Spec('mpileaks')
-        perms = spack.package_prefs.get_package_permissions(spec)
+        perms = spp.get_package_permissions(spec)
         assert perms == stat.S_IRWXU | stat.S_IRGRP | stat.S_IXGRP
 
-        dir_perms = spack.package_prefs.get_package_dir_permissions(spec)
+        dir_perms = spp.get_package_dir_permissions(spec)
         expected = stat.S_IRWXU | stat.S_IRGRP | stat.S_IXGRP | stat.S_ISGID
         assert dir_perms == expected
 
-        group = spack.package_prefs.get_package_group(spec)
+        group = spp.get_package_group(spec)
         assert group == 'mpileaks'
 
     def test_config_perms_fail_write_gt_read(self, configure_permissions):
         # Test failure for writable more permissive than readable
         spec = Spec('callpath')
         with pytest.raises(ConfigError):
-            spack.package_prefs.get_package_permissions(spec)
+            spp.get_package_permissions(spec)
